requirements.txt
tensorflow>=2.13.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=9.0.0

